--- a/guardrails/constraint_generator/__init__.py
+++ b/guardrails/constraint_generator/__init__.py
@@ -1,0 +1,6 @@
+from guardrails.constraint_generator.constraint_generator import ConstraintGenerator
+from guardrails.constraint_generator.balanced_braces_generator import (
+    BalancedBracesGenerator,
+)
+
+__all__ = ["BalancedBracesGenerator", "ConstraintGenerator"]