--- a/guardrails/constraint_generator/__init__.py
+++ b/guardrails/constraint_generator/__init__.py
@@ -2,5 +2,18 @@ from guardrails.constraint_generator.constraint_generator import ConstraintGener
 from guardrails.constraint_generator.balanced_braces_generator import (
     BalancedBracesGenerator,
 )
+from guardrails.constraint_generator.json_constraint_generator import (
+    JSONConstraintGenerator,
+    KeywordConstraintGenerator,  # JC Note: Do we want to expose this?
+    NumberConstraintGenerator,
+    UnionConstraintGenerator,
+)
 
-__all__ = ["BalancedBracesGenerator", "ConstraintGenerator"]
+__all__ = [
+    "BalancedBracesGenerator",
+    "ConstraintGenerator",
+    "JSONConstraintGenerator",
+    "KeywordConstraintGenerator",
+    "NumberConstraintGenerator",
+    "UnionConstraintGenerator",
+]