--- a/guardrails/constraint_generator/__init__.py
+++ b/guardrails/constraint_generator/__init__.py
@@ -4,8 +4,10 @@ from guardrails.constraint_generator.balanced_braces_generator import (
 )
 from guardrails.constraint_generator.json_constraint_generator import (
     JSONConstraintGenerator,
+    JSONValueConstraint,
     KeywordConstraintGenerator,  # JC Note: Do we want to expose this?
     NumberConstraintGenerator,
+    QuotedStringConstraintGenerator,
     UnionConstraintGenerator,
 )
 
@@ -13,7 +15,9 @@ __all__ = [
     "BalancedBracesGenerator",
     "ConstraintGenerator",
     "JSONConstraintGenerator",
+    "JSONValueConstraint",
     "KeywordConstraintGenerator",
     "NumberConstraintGenerator",
+    "QuotedStringConstraintGenerator",
     "UnionConstraintGenerator",
 ]