--- a/guardrails/constraint_generator/balanced_braces_generator.py
+++ b/guardrails/constraint_generator/balanced_braces_generator.py
@@ -4,14 +4,16 @@ from guardrails.constraint_generator import ConstraintGenerator
 
 
 class BalancedBracesGenerator(ConstraintGenerator):
-    def __init__(self, max_depth: int = 1):
+    def __init__(self, max_depth: int):
         self.max_depth = max_depth
         self.current_depth = 0
 
     def get_valid_tokens(self) -> Optional[Set[str]]:
         if self.current_depth < 0:
             return set()  # We have closed more than we opened.
-        if self.current_depth >= self.max_depth:
+        elif self.current_depth == 0:
+            return {"{"}
+        elif self.max_depth != 0 and self.current_depth >= self.max_depth:
             return {"}"}
         else:
             return {"{", "}"}