--- a/guardrails/formatters/base_formatter.py
+++ b/guardrails/formatters/base_formatter.py
@@ -7,6 +7,10 @@ from guardrails.llm_providers import (
 
 
 class BaseFormatter(ABC):
+    """A Formatter takes an LLM Callable and wraps the method into an abstract callable.
+    Used to perform manipulations of the input or the output, like JSON constrained-
+    decoding."""
+
     @abstractmethod
     def wrap_callable(self, llm_callable: PromptCallableBase) -> ArbitraryCallable: ...